Move PluginManager to its own core module
